requirements.txt
pandas
numpy 
setuptools
google-cloud-storage
pyyaml
tensorflow
matplotlib
joblib